moved get_dataset_capacilities_url under helpers
